perf(web): extract React Query key factories to a leaf-level module
`useWebSocket` is mounted at the app shell, runs on every authenticated
visit, and only needs the `*Keys` factories to invalidate caches when
events arrive. Previously each domain hook (`useTasks`, `useTimer`,
`useTimeBlocks`, `useSubtasks`, `useCalendars`) owned + exported its own
key factory, so importing `taskKeys` from `useTasks` dragged 940 lines
of mutation/state code into `useWebSocket`'s static graph — and from
there into the boot bundle.

Pulled all five key factories into a single `lib/query-keys.ts`. Each
domain hook re-exports its own key from this file for backwards compat,
so consumers like `useTasks.taskKeys` keep working without changes.

Result: entry chunk drops from 158 kB → 143 kB. The mutation code
stays behind whatever route or component actually uses it.
